Extract contextual template UI module and clean imports
